refactor: button functions
