+++ src/utils/buttonFunctions.js
@@ -1,0 +1,13 @@
+const handleKeepBuyingBtn = () => {
+  alert('Continue comprando!');
+};
+
+const handlePaymentBtn = () => {
+  alert('Redirecionando para o pagamento!');
+};
+
+const handleBagBtn = () => {
+  alert('Adicionando à sacola!');
+};
+
+export { handleKeepBuyingBtn, handlePaymentBtn, handleBagBtn };
